requirements.txt
# --- Core Framework ---
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.1

# --- Database (PostgreSQL) ---
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# --- Validation & Settings ---
pydantic==2.5.3
pydantic-settings==2.1.0

# --- Authentication ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1

# --- HTTP Client ---
httpx==0.26.0

# --- Utilities ---
fastapi-mail==1.4.1
xhtml2pdf==0.2.14

# --- Testing ---
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0